task 3 fix 1

+++ python-exceptions/3-safe_print_division.py
@@ -2,7 +2,7 @@
 
 def safe_print_division(a, b):
     try: 
-        result = int(a / b)
+        result = a / b
     except ArithmeticError:
         result = "None"
     finally:
